+++ tests/testthat/test-plot_graph_3d.R
@@ -5,8 +5,7 @@ test_that("plot_graph_3d works", {
   plt <- suppressWarnings(
     KGExplorer::plot_graph_3d(g=g, show_plot = FALSE)
   )
-
   testthat::expect_true(methods::is(g,"igraph"))
-  testthat::expect_length(g,34)
+  testthat::expect_length(g,23)
   testthat::expect_true(methods::is(plt,"plotly"))
 })